+++ pkg/loki/config_wrapper.go
@@ -31,6 +31,7 @@ type ConfigWrapper struct {
 	PrintVersion    bool
 	VerifyConfig    bool
 	PrintConfig     bool
+	ListTargets     bool
 	LogConfig       bool
 	ConfigFile      string
 	ConfigExpandEnv bool
@@ -40,6 +41,7 @@ func (c *ConfigWrapper) RegisterFlags(f *flag.FlagSet) {
 	f.BoolVar(&c.PrintVersion, "version", false, "Print this builds version information")
 	f.BoolVar(&c.VerifyConfig, "verify-config", false, "Verify config file and exits")
 	f.BoolVar(&c.PrintConfig, "print-config-stderr", false, "Dump the entire Loki config object to stderr")
+	f.BoolVar(&c.ListTargets, "list-targets", false, "List available targets")
 	f.BoolVar(&c.LogConfig, "log-config-reverse-order", false, "Dump the entire Loki config object at Info log "+
 		"level with the order reversed, reversing the order makes viewing the entries easier in Grafana.")
 	f.StringVar(&c.ConfigFile, "config.file", "", "yaml file to load")